requirements.txt
anyio==3.5.0
asgiref==3.5.1
click==8.1.3
fastapi==0.77.1
h11==0.13.0
idna==3.3
pydantic==1.9.0
python-magic==0.4.25
sniffio==1.2.0
starlette==0.19.1
typing-extensions==4.2.0
uvicorn==0.17.6
